[ELF] Lookup INPUT argument in the current directory
If an argument of the INPUT directive is a regular path, linker should
lookup it in the current folder first.

The fix does not contain any test cases because I think it is not a good
idea to pollute a current folder (which in general might be arbitrary)
by test files.

Differential Revision: http://reviews.llvm.org/D15027

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@254178 91177308-0d34-0410-b5e6-96231b3b80d8