Verify that macho-o delta64 relocs have the same offset.
The delta64 relocation is represented as the pair ARM64_RELOC_SUBTRACTOR and ARM64_RELOC_UNSIGNED.

Those should always have the same offset, so this adds a check and tests to ensure this is the case.

Also updated the error printing in this case to shows both relocs when erroring on pair.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@255274 91177308-0d34-0410-b5e6-96231b3b80d8