Merge identical strings.
This avoids the need to have reserve and addString in sync.

We avoid hashing the global symbols again. This means that we don't
merge a global symbol that has the same name as some other string, but
that doesn't seem very common. The string table size is the same in
clang an scylladb with or without hashing global symbols again.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@259136 91177308-0d34-0410-b5e6-96231b3b80d8