Move SimpleDefinedAtom::addReference to DefinedAtom.
Most of the other methods to access Reference's were on DefinedAtom so
this just keeps them all together.

This will be used in a future patch in ObjCPass which needs to add
new references.  The method is virtual because we may add references to
different data structures depending on whether we parsed a macho file or
yaml.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@259436 91177308-0d34-0410-b5e6-96231b3b80d8