Add support for -sdk_version cmdline option.
This option is emitted in the min_version load commands.

Note, there's currently a difference in behaviour compared to ld64 in
that we emit a warning if we generate a min_version load command and
didn't give an sdk_version.  We need to decide what the correct behaviour
is here as its possible we want to emit an error and force clients to
provide the option.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@259729 91177308-0d34-0410-b5e6-96231b3b80d8