ELF: Create LinkerScript class to move code out of Writer.
Previously, we had code for linker scripts in Writer. This patch
separates that as LinkerScript class. The class provides a few
functions to query linker scripts and is also a container of some
linker-script-specific information.

Hopefully, Writer will only implement the default behavior and let
the new class handle gotchas regarding linker scripts.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@260591 91177308-0d34-0410-b5e6-96231b3b80d8