Fix some confusion about what can be preempted.
For shared libraries we allow any weak undefined symbol to eventually be
resolved, even if we never see a definition in another .so. This matches
the behavior when handling other undefined symbols in a shared library.

For executables, we require seeing a definition in a .so or resolve it
to zero. This is also similar to how non weak symbols are handled.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@262017 91177308-0d34-0410-b5e6-96231b3b80d8