Delete isTlsDynRel.
It was a badly specified hack for when a tls relocation should be
propagated to the dynamic relocation table.

This replaces it with a not as bad hack of saying that a local dynamic
tls relocation is never preempted.

I will try to remove even that second hack in the next patch.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@262955 91177308-0d34-0410-b5e6-96231b3b80d8