Always pass an allocator to YAMLTraits.
The YAML traits new's when not passed an allocator to parse data.
For atom types, this is a leak as we don't destruct atoms.  For
the File here, we do actually destruct File's so that single case of
not using an allocator will be fine.

Should fix a bunch more leaks.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@263680 91177308-0d34-0410-b5e6-96231b3b80d8