[COFF] Don't call memcpy with a NULL argument
Some declarations of memcpy (like glibc's for example) are attributed
with notnull which makes it UB for NULL to get passed in, even if the
memcpy count is zero.

To account for this, guard the memcpy with an appropriate precondition.

This should fix the last UBSan bug, exposed by the test suite, in the
COFF linker.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@263919 91177308-0d34-0410-b5e6-96231b3b80d8