Skip some relocations in scanRelocs.
When a tls access is optimized, a group of relocations is converted at a
time.

We were already skipping relocations that were optimized out in
relocate, but not in scanRelocs.

This is a small optimization. I got here while working on a patch that
will always keep scanRelocs and relocate in sync.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@264048 91177308-0d34-0410-b5e6-96231b3b80d8