[LTO] Keep linkonce_odr symbols when appropriate.
Ensure we keep the symbol we need to before it reaches
the Writer (and hit an assertion), changing its linkage
from linkonce_odr to weak. For a more detailed description
of the problem, see PR19901 where a similar problem was
fixed for the gold plugin. Thanks to Rafael for providing
a testcase.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@264111 91177308-0d34-0410-b5e6-96231b3b80d8