Use fewer allocators.
Instead of having 3 section allocators per file, have 3 for all files.

This is a substantial performance improvement for some cases. Linking
chromium without gc speeds up by 1.065x.

This requires using _exit in fatal since we have to avoid destructing
an InputSection if fatal is called from the constructor.

Thanks to Rui for the suggestion.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@285290 91177308-0d34-0410-b5e6-96231b3b80d8