Delay computation of IsAbsolute.
We parse linker scripts very early, but whether an expression is
absolute or not can depend on a symbol defined in a .o. Given that, we
have to delay the computation of IsAbsolute. We can do that by storing
an AST when parsing or by also making IsAbsolute a function like we do
for the expression value. This patch implements the second option.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@285628 91177308-0d34-0410-b5e6-96231b3b80d8