Rewrite CommonInputSection as a synthetic input section.
A CommonInputSection is a section containing all common symbols.
That was an input section but was abstracted in a different way
than the synthetic input sections because it was written before
the synthetic input section was invented.

This patch rewrites CommonInputSection as a synthetic input section
so that it behaves better with other sections.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@286053 91177308-0d34-0410-b5e6-96231b3b80d8