build: add support for standalone lld build
Enable building lld as a standalone project. This is motivated by the desire to
package lld for inclusion in a linux distribution. This allows building lld
against an existing paired llvm installation. Now that lld is usable on x86_64,
it makes sense to revive this configuration to allow distributions to package
it.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@289421 91177308-0d34-0410-b5e6-96231b3b80d8