ELF: Place relro sections after non-relro sections in r/w segment.
This is in preparation for my next change, which will introduce a relro
nobits section. That requires that relro sections appear at the end of the
progbits part of the r/w segment so that the relro nobits section can appear
contiguously.

Because of the amount of churn required in the test suite, I'm making this
change separately.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@291523 91177308-0d34-0410-b5e6-96231b3b80d8