[test] Fix zlib cond when building stand-alone, clean up
Fix the test zlib conditional to use LLVM_ENABLE_ZLIB value when
building stand-alone. The HAVE_LIBZ is not available when performing
a stand-alone build. Since the zlib support is a feature of
the underlying LLVM library, it exports the actual status as the final
value of LLVM_ENABLE_ZLIB in LLVMConfig.

While at it, canonicalize the boolean value into 0/1 and remove unused
CMake definitions (most likely copied from clang).

Differential Revision: https://reviews.llvm.org/D29340

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@294508 91177308-0d34-0410-b5e6-96231b3b80d8