Handle .eh_frame pointing to discarded section in -r.
This is a really horrible case. If a .eh_frame points to a discarded
section, it is not clear what is the correct thing to do.

It looks like ld.bfd discards the entire .eh_frame content and gold
discards the second relocation, leaving one frame with an fde that
refers to a bogus location. This is similar to what gold does.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@295133 91177308-0d34-0410-b5e6-96231b3b80d8